finish splitting files for datatypes
