+++ src/table/Table.ts
@@ -1,0 +1,193 @@
+/* *****************************************************************************
+ * Caleydo - Visualization for Molecular Biology - http://caleydo.org
+ * Copyright (c) The Caleydo Team. All rights reserved.
+ * Licensed under the new BSD license, available at http://caleydo.org/license
+ **************************************************************************** */
+/**
+ * Created by Samuel Gratzl on 04.08.2014.
+ */
+
+import {IPersistable, mixin} from '../index';
+import {Range, all, list as rlist, parse, RangeLike} from '../range';
+import {resolve as idtypes_resolve, createLocalAssigner} from '../idtype';
+import {IValueType, VALUE_TYPE_INT, VALUE_TYPE_REAL, guessValueTypeDesc} from '../datatype';
+import {ITable, ITableColumn, ITableDataDescription, createDefaultTableDesc} from './ITable';
+import ATable from './ATable';
+import TableVector from './internal/TableVector';
+import {ITableLoader, ITableLoader2, adapterOne2Two, viaAPI2Loader, viaDataLoader} from './loader';
+
+/**
+ * root matrix implementation holding the data
+ */
+export default class Table extends ATable implements ITable {
+  private vectors: TableVector[];
+
+  constructor(public readonly desc: ITableDataDescription, private loader: ITableLoader2) {
+    super(null);
+    this.root = this;
+    this.vectors = desc.columns.map((cdesc, i) => new TableVector(this, i, cdesc));
+  }
+
+  get idtype() {
+    return idtypes_resolve(this.desc.idtype || (<any>this.desc).rowtype);
+  }
+
+  get idtypes() {
+    return [this.idtype];
+  }
+
+  col(i: number) {
+    return this.vectors[i];
+  }
+
+  cols(range: RangeLike = all()) {
+    return parse(range).filter(this.vectors, [this.ncol]);
+  }
+
+  /**
+   * access at a specific position
+   * @param i
+   * @param j
+   * @returns {*}
+   */
+  at(i: number, j: number) {
+    return this.colData(this.col(j).column, rlist(i)).then((arr) => arr[0]);
+  }
+
+  queryView(name: string, args: any) {
+    return new Table(this.desc, adapterOne2Two(this.loader.view(this.desc, name, args)));
+  }
+
+  data(range: RangeLike = all()) {
+    return this.loader.data(this.desc, parse(range));
+  }
+
+  colData(column: string, range: RangeLike = all()) {
+    return this.loader.col(this.desc, column, parse(range));
+  }
+
+  objects(range: RangeLike = all()) {
+    return this.loader.objs(this.desc, parse(range));
+  }
+
+  /**
+   * return the row ids of the matrix
+   * @returns {*}
+   */
+  rows(range: RangeLike = all()): Promise<string[]> {
+    return this.loader.rows(this.desc, parse(range));
+  }
+
+  rowIds(range: RangeLike = all()) {
+    return this.loader.rowIds(this.desc, parse(range));
+  }
+
+  ids(range: RangeLike = all()) {
+    return this.rowIds(range);
+  }
+
+  size() {
+    return this.desc.size;
+  }
+
+  persist() {
+    return this.desc.id;
+  }
+
+  restore(persisted: any): IPersistable {
+    if (persisted && typeof persisted.col === 'number') {
+      return this.col(persisted.col);
+    }
+    return super.restore(persisted);
+  }
+}
+
+
+/**
+ * module entry point for creating a datatype
+ * @param desc
+ * @returns {ITable}
+ */
+export function create(desc: ITableDataDescription, loader?: ITableLoader): ITable {
+  if (loader) {
+    return new Table(desc, adapterOne2Two(loader));
+  }
+  return new Table(desc, viaAPI2Loader());
+}
+
+export function wrapObjects(desc: ITableDataDescription, data: any[], nameProperty: string|((obj: any) => string)) {
+  return new Table(desc, adapterOne2Two(viaDataLoader(data, nameProperty)));
+}
+
+function toObjects(data: any[][], cols: string[]) {
+  return data.map((row) => {
+    const r: any = {};
+    cols.forEach((col, i) => r[col] = row[i]);
+    return r;
+  });
+}
+function toList(objs: any[], cols: string[]) {
+  return objs.map((obj) => cols.map((c) => obj[c]));
+}
+
+export interface IAsTableOptions {
+  name?: string;
+  idtype?: string;
+  rowassigner?(ids: string[]): Range;
+  keyProperty?: string;
+}
+
+
+function asTableImpl(columns: ITableColumn[], rows: string[], objs: any[], data: IValueType[][], options: IAsTableOptions = {}) {
+  const desc = mixin(createDefaultTableDesc(), {
+    columns: columns,
+    size: [rows.length, columns.length]
+  }, options);
+
+  const rowAssigner = options.rowassigner || createLocalAssigner();
+  const loader: ITableLoader = () => {
+    const r = {
+      rowIds: rowAssigner(rows),
+      rows: rows,
+      objs: objs,
+      data: data
+    };
+    return Promise.resolve(r);
+  };
+  return new Table(desc, adapterOne2Two(loader));
+}
+
+export function asTableFromArray(data: any[][], options: IAsTableOptions = {}): ITable {
+  const rows = data.map((r) => r[0]);
+  const cols = data[0].slice(1);
+  const tableData = data.slice(1).map((r) => r.slice(1));
+
+  const columns = cols.map((col, i) => {
+    return {
+      name: col,
+      value: guessValueTypeDesc(tableData.map((row) => row[i]))
+    };
+  });
+
+  const realData = tableData.map((row) => columns.map((col, i) => (col.value.type === VALUE_TYPE_REAL || col.value.type === VALUE_TYPE_INT) ? parseFloat(row[i]) : row[i]));
+  const objs = toObjects(realData, cols);
+
+  return asTableImpl(columns, rows, objs, realData, options);
+}
+
+export function asTable(data: any[], options: IAsTableOptions = {}): ITable {
+  const keyProperty = options.keyProperty || '_id';
+
+  const rows = data.map((r, i) => String(r[keyProperty]) || String(i));
+  const cols = Object.keys(data[0]);
+  const objs = data;
+  const realData = toList(objs, cols);
+
+  const columns = cols.map((col, i) => {
+    return {
+      name: col,
+      value: guessValueTypeDesc(realData.map((row) => row[i]))
+    };
+  });
+  return asTableImpl(columns, rows, objs, realData, options);
+}
