+++ src/table/index.ts
@@ -2,4 +2,14 @@
  * Created by sam on 26.12.2016.
  */
 
+import {IVector} from '../vector';
+import VectorTable from './internal/VectorTable';
+import {IDataDescription} from '../datatype';
+
+
 export {ITable, IQueryArgs, ITableColumn, ITableDataDescription} from './ITable';
+export {asTable, asTableFromArray, IAsTableOptions} from './Table';
+
+export function fromVectors(desc: IDataDescription, vecs: IVector[]) {
+  return new VectorTable(desc, vecs);
+}
